refactor(types): Props 관련 타입 분리
