+++ src/components/layout/Layout.tsx
@@ -1,11 +1,7 @@
-import { ReactNode } from "react";
 import { Header } from "./Header";
 import { Sidebar } from "./Sidebar";
 import { Footer } from "./Footer";
-
-interface LayoutProps {
-    children: ReactNode;
-}
+import { LayoutProps } from "@/types/props";
 
 export const Layout = ({ children }: LayoutProps) => {
     return (
