--- src/types/sidebar.ts
@@ -1,6 +1,0 @@
-export interface SideBarItem {
-    id: string;
-    name: string;
-    href: string;
-    icon: string; // 이미지 경로
-}
